+++ src/Qux/Commands/Check.hs
@@ -11,6 +11,11 @@ Maintainer  : public@hjwylde.com
 
 module Qux.Commands.Check where
 
+import Control.Monad.Except
+
+import Language.Qux.Annotated.Parser
+import Language.Qux.Annotated.Syntax
+
 import qualified Qux.Commands.Build as Build
 
 
@@ -19,7 +24,13 @@ data Options = Options {
     }
 
 handle :: Options -> IO ()
-handle options = Build.handle Build.Options {
+handle options = Build.handle $ buildOptions options
+
+check :: Program SourcePos -> Except String ()
+check = Build.build $ buildOptions undefined
+
+buildOptions :: Options -> Build.Options
+buildOptions options = Build.Options {
     Build.optTypeCheck = True,
     Build.argFilePaths = argFilePaths options
     }