+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 * Added `--numeric-version` and `--qux-version` global options.
 * Added ability to parse any valid Qux value as an argument to `qux run`.
 * Added ability to build and check multiple files at once.
-* Updated build and check to print multiple errors
+* Updated build and check to print multiple errors.
+* Added program checking as default to `qux run' and option to skip it, `--skip-checks'.
 
 #### v0.1.0.0